+++ src/xkbcomp/keycodes.c
@@ -94,7 +94,7 @@ AddLedName(KeyNamesInfo *info, bool same_file,
             xkb_led_index_t use    = (replace ? new_idx + 1 : old_idx + 1);
             xkb_led_index_t ignore = (replace ? old_idx + 1 : new_idx + 1);
             log_warn(info->ctx, XKB_LOG_MESSAGE_NO_ID,
-                     "Multiple indicators named %s; Using %d, ignoring %d\n",
+                     "Multiple indicators named %s; Using %u, ignoring %u\n",
                      xkb_atom_text(info->ctx, new->name), use, ignore);
         }
 
@@ -116,7 +116,7 @@ AddLedName(KeyNamesInfo *info, bool same_file,
             const xkb_atom_t use    = (replace ? new->name : old->name);
             const xkb_atom_t ignore = (replace ? old->name : new->name);
             log_warn(info->ctx, XKB_LOG_MESSAGE_NO_ID,
-                     "Multiple names for indicator %d; "
+                     "Multiple names for indicator %u; "
                      "Using %s, ignoring %s\n", new_idx + 1,
                      xkb_atom_text(info->ctx, use),
                      xkb_atom_text(info->ctx, ignore));
@@ -417,8 +417,8 @@ HandleKeycodeDef(KeyNamesInfo *info, KeycodeDef *stmt, bool report)
 {
     if (stmt->value < 0 || stmt->value > XKB_KEYCODE_MAX) {
         log_err(info->ctx, XKB_LOG_MESSAGE_NO_ID,
-                "Illegal keycode %lld: must be between 0..%u; "
-                "Key ignored\n", (long long) stmt->value, XKB_KEYCODE_MAX);
+                "Illegal keycode %"PRId64": must be between 0..%u; "
+                "Key ignored\n", stmt->value, XKB_KEYCODE_MAX);
         return false;
     }
